Fixed composition of receive handlers
